♻️ Refactor: formatDate 함수 util로 분리
